# api-gateway/src/test/resources/openapi/growbusiness-gateway-api.yaml
openapi: 3.0.3
info:
  title: GrowBusiness API Gateway
  description: |
    Semua endpoint microservices RSU Sumatera melalui **API Gateway** (`http://localhost:8000`).

    **Alur testing:**
    1. `POST /api/auth/login` dengan JSON (contoh admin di bawah)
    2. Salin `token` dari response
    3. Klik **Authorize** → masukkan `Bearer <token>`
    4. Panggil endpoint lain dengan body JSON sesuai contoh

    **Role:** ADMIN, DOCTOR, PHARMACIST, STAFF, USER
  version: 1.0.0
  contact:
    name: GrowBusiness RS
servers:
  - url: http://localhost:8000
    description: API Gateway (local)
...
security:
  - bearerAuth: []
paths:
...
  /api/pharmacy/medicines:
    post:
      tags: [Medicines]
      summary: Create medicine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObatRequest'
            example:
              kodeObat: OB001
              namaObat: Paracetamol 500mg
              deskripsi: Analgesik dan antipiretik
              harga: 5000
              satuan: tablet
              stok: 100
              stokMinimal: 20
              supplier: PT Kimia Farma
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseObat'
    get:
      tags: [Medicines]
      summary: Get all medicines
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponseObat'
  /api/pharmacy/medicines/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Medicines]
      summary: Get medicine by ID
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseObat'
  /api/pharmacy/medicines/search:
    get:
      tags: [Medicines]
      summary: Search medicines by name
      parameters:
        - name: nama
          in: query
          required: true
          schema:
            type: string
          example: Paracetamol
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponseObat'
...
  /api/pharmacy/medicines/{id}/stock:
    parameters:
      - $ref: '#/components/parameters/IdPath'
      - name: stok
        in: query
        required: true
        schema:
          type: integer
        example: 150
    patch:
      tags: [Medicines]
      summary: Update medicine stock
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseObat'
  /api/payment/invoices:
    post:
      tags: [Invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagihanCreateRequest'
            example:
              pasienId: 1
              biayaKonsultasi: 150000
              hargaObat: 50000
              diskonAsuransi: 20000
      responses:
        '201':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseTagihan'
              example:
                message: Invoice created successfully
                data:
                  id: 1
                  noInvoice: INV-ABC123DE
                  pasienId: 1
                  biayaKonsultasi: 150000
                  hargaObat: 50000
                  subtotal: 200000
                  diskonAsuransi: 20000
                  pajakPPN: 18000
                  totalBayar: 198000
                  status: PENDING
  /api/payment/invoices/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Invoices]
      summary: Get invoice by ID
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseTagihan'
  /api/payment/invoices/patient/{pasienId}:
    get:
      tags: [Invoices]
      summary: Get invoices for patient
      parameters:
        - name: pasienId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponseTagihan'
...
  /api/payment/invoices/prescription/{resepId}:
    get:
      tags: [Invoices]
      summary: Get invoice by prescription ID
      parameters:
        - name: resepId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseTagihan'
  /api/payment/invoices/{id}/status:
    parameters:
      - $ref: '#/components/parameters/IdPath'
      - name: status
        in: query
        required: true
        schema:
          type: string
          example: PAID
    patch:
      tags: [Invoices]
      summary: Update invoice status (ADMIN)
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseTagihan'
  /api/payment/invoices/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [Invoices]
      summary: Cancel invoice (saga compensation)
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseTagihan'
  /api/payment/transactions/process:
    post:
      tags: [Transactions]
      summary: Process payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransaksiRequest'
            example:
              tagihanId: 1
              jumlahBayar: 198000
              metodePembayaran: TUNAI
              referensiTransaksi: TRX20260515001
      responses:
        '201':
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseTransaksi'
              example:
                message: Payment processed successfully
                data:
                  id: 1
                  tagihanId: 1
                  jumlahBayar: 198000
                  metodePembayaran: TUNAI
                  statusPembayaran: COMPLETED
                  referensiTransaksi: TRX20260515001
  /api/payment/transactions/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Transactions]
      summary: Get transaction by ID
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponseTransaksi'
  /api/payment/transactions/invoice/{tagihanId}:
    get:
      tags: [Transactions]
      summary: Get transactions for invoice
      parameters:
        - name: tagihanId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponseTransaksi'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Token dari POST /api/auth/login. Format Authorize: Bearer <token>'
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      example: 1
...
  schemas:
...
    ObatRequest:
      type: object
      required: [kodeObat, namaObat]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        kodeObat:
          type: string
        namaObat:
          type: string
        deskripsi:
          type: string
        harga:
          type: number
          format: double
        satuan:
          type: string
        stok:
          type: integer
        stokMinimal:
          type: integer
        supplier:
          type: string
    TagihanCreateRequest:
      type: object
      required: [pasienId]
      properties:
        pasienId:
          type: integer
          format: int64
        biayaKonsultasi:
          type: number
          format: double
        hargaObat:
          type: number
          format: double
        diskonAsuransi:
          type: number
          format: double
        resepId:
          type: integer
          format: int64
        namaObat:
          type: string
        jumlahObat:
          type: integer
    Tagihan:
      type: object
      properties:
        id:
          type: integer
          format: int64
        pasienId:
          type: integer
          format: int64
        resepId:
          type: integer
          format: int64
        noInvoice:
          type: string
        biayaKonsultasi:
          type: number
        hargaObat:
          type: number
        subtotal:
          type: number
        diskonAsuransi:
          type: number
        pajakPPN:
          type: number
        totalBayar:
          type: number
        status:
          type: string
    TransaksiRequest:
      type: object
      required: [tagihanId, jumlahBayar, metodePembayaran]
      properties:
        tagihanId:
          type: integer
          format: int64
        jumlahBayar:
          type: number
          format: double
        metodePembayaran:
          type: string
          example: TUNAI
        referensiTransaksi:
          type: string
        keterangan:
          type: string
    Transaksi:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tagihanId:
          type: integer
          format: int64
        jumlahBayar:
          type: number
        metodePembayaran:
          type: string
        statusPembayaran:
          type: string
        referensiTransaksi:
          type: string
        keterangan:
          type: string
...
    DataResponseObat:
      type: object
      properties:
        message:
          type: string
        data:
          $ref: '#/components/schemas/ObatRequest'
    ListResponseObat:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ObatRequest'
        count:
          type: integer
    DataResponseTagihan:
      type: object
      properties:
        message:
          type: string
        data:
          $ref: '#/components/schemas/Tagihan'
    ListResponseTagihan:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Tagihan'
        count:
          type: integer
    DataResponseTransaksi:
      type: object
      properties:
        message:
          type: string
        data:
          $ref: '#/components/schemas/Transaksi'
    ListResponseTransaksi:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Transaksi'
        count:
          type: integer
...